[refect] Separated Files and Functions by Feature
