+++ includes/miniRT.h
@@ -6,7 +6,7 @@
 /*   By: phan <phan@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/04 11:57:07 by hcho2             #+#    #+#             */
-/*   Updated: 2023/09/16 18:56:47 by phan             ###   ########.fr       */
+/*   Updated: 2023/09/19 15:12:06 by phan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
